requirements.txt
Django==4.1.3
django-environ==0.9.0
django-phonenumber-field[phonenumbers]==7.0.1
mysqlclient==2.1.1
pre-commit==2.20.0
